+++ index.js
@@ -14,7 +14,7 @@
 
 'use strict'
 
-export { default, default as LifiSwidgeProtocol } from './src/lifi-swidge-protocol.js'
+export { default, default as LifiSwidgeProtocol, NATIVE_VALUE_BRIDGE_DENY_LIST } from './src/lifi-swidge-protocol.js'
 
 /** @typedef {import('./src/lifi-swidge-protocol.js').LifiSwidgeProtocolConfig} LifiSwidgeProtocolConfig */
 /** @typedef {import('./src/lifi-swidge-protocol.js').LifiRouteOrder} LifiRouteOrder */